+++ desktop/CHANGELOG.md
@@ -20,6 +20,7 @@
 ### Bug Fixes
 
 - fixed `ssb-profile` not allowing `headerImage` and `avatarImage` to be set to null
+- fixed bug where you could no longer see profile information on a group application
 
 ## 1.5.1